Changed Display, add Delete, save Notes
Changed the Way the apps looks when trying to add/ edit a note.

Added the Delete Button to give user access to delete a note

added the ability to save a note and go back to it later, previously
when pressing save and opening the app again it will be a new Note.
When app is closed, the notes will be gone.